Engine: move partialTransaction to interpreter
CHANGELOG_BEGIN
CHANGELOG_END
