+++ daml-lf/transaction/src/main/scala/com/digitalasset/daml/lf/transaction/Transaction.scala
@@ -9,11 +9,11 @@ import com.digitalasset.daml.lf.data._
 import com.digitalasset.daml.lf.language.LanguageVersion
 import com.digitalasset.daml.lf.transaction.Node._
 import com.digitalasset.daml.lf.value.Value
-import com.digitalasset.daml.lf.value.Value._
+
 import scalaz.Equal
 
 import scala.annotation.tailrec
-import scala.collection.breakOut
+
 import scala.collection.immutable.HashMap
 
 case class VersionedTransaction[Nid, Cid](
@@ -69,10 +69,6 @@ case class VersionedTransaction[Nid, Cid](
   *                     This is a hint for what packages are required to validate
   *                     the transaction using the current interpreter.
   *                     The used packages are not serialized using [[TransactionCoder]].
-  * @param transactionSeed master hash used to derived node and relative contractId
-  *                        discriminators. If it is undefined, the discriminators have not be
-  *                        generated and have be let undefined in the nodes and the relative
-  *                        contractIds of the transaction.
   *
   * Users of this class may assume that all instances are well-formed, i.e., `isWellFormed.isEmpty`.
   * For performance reasons, users are not required to call `isWellFormed`.
@@ -392,415 +388,5 @@ object Transaction {
     */
   final case class ContractNotActive(coid: TContractId, templateId: TypeConName, consumedBy: NodeId)
       extends TransactionError
-
-  object PartialTransaction {
-
-    type NodeIdx = Value.NodeIdx
-
-    /** Contexts of the transaction graph builder, which we use to record
-      * the sub-transaction structure due to 'exercises' statements.
-      */
-    sealed abstract class Context extends Product with Serializable {
-      def contextSeed: Option[crypto.Hash]
-
-      def children: BackStack[NodeId]
-
-      def addChild(child: NodeId): Context
-    }
-
-    /** The root context, which is what we use when we are not exercising
-      * a choice.
-      */
-    final case class ContextRoot(
-        val contextSeed: Option[crypto.Hash],
-        children: BackStack[NodeId] = BackStack.empty,
-    ) extends Context {
-      override def addChild(child: NodeId): ContextRoot = copy(children = children :+ child)
-    }
-
-    /** Context when creating a sub-transaction due to an exercises. */
-    final case class ContextExercise(
-        ctx: ExercisesContext,
-        children: BackStack[NodeId] = BackStack.empty,
-    ) extends Context {
-      override def addChild(child: NodeId): ContextExercise =
-        copy(children = children :+ child)
-
-      override def contextSeed: Option[crypto.Hash] = ctx.contextSeed
-    }
-
-    /** Context information to remember when building a sub-transaction
-      *  due to an 'exercises' statement.
-      *
-      *  @param targetId Contract-id referencing the contract-instance on
-      *                  which we are exercising a choice.
-      *  @param templateId Template-id referencing the template of the
-      *                    contract on which we are exercising a choice.
-      *  @param contractKey Optional contract key, if defined for the
-      *                     contract on which we are exercising a choice.
-      *  @param choiceId Label of the choice that we are exercising.
-      *  @param consuming True if the choice consumes the contract.
-      *  @param actingParties The parties exercising the choice.
-      *  @param chosenValue The chosen value.
-      *  @param signatories The signatories of the contract.
-      *  @param stakeholders The stakeholders of the contract.
-      *  @param controllers The controllers of the choice.
-      *  @param nodeId The node to be inserted once we've
-      *                         finished this sub-transaction.
-      *  @param parent The context in which the exercises is
-      *                       happening.
-      */
-    case class ExercisesContext(
-        contextSeed: Option[crypto.Hash],
-        targetId: TContractId,
-        templateId: TypeConName,
-        contractKey: Option[KeyWithMaintainers[Value[Nothing]]],
-        choiceId: ChoiceName,
-        optLocation: Option[Location],
-        consuming: Boolean,
-        actingParties: Set[Party],
-        chosenValue: Value[TContractId],
-        signatories: Set[Party],
-        stakeholders: Set[Party],
-        controllers: Set[Party],
-        nodeId: NodeId,
-        parent: Context,
-    )
-
-    /** The initial transaction from which we start building. It does not
-      *  contain any nodes and is not marked as aborted.
-      *  @param transactionSeed is the master hash used to derive nodes and
-      *                         contractIds discriminator. If let undefined no
-      *                         discriminators should be generate.
-      */
-    def initial(seedWithTime: Option[(crypto.Hash, Time.Timestamp)] = None) =
-      PartialTransaction(
-        seedWithTime.map(_._2),
-        nextNodeIdx = 0,
-        nodes = HashMap.empty,
-        consumedBy = Map.empty,
-        context = ContextRoot(seedWithTime.map(_._1)),
-        aborted = None,
-        keys = Map.empty,
-      )
-
-  }
-
-  /** A transaction under construction
-    *
-    *  @param nodes The nodes of the transaction graph being built up.
-    *  @param consumedBy 'ContractId's of all contracts that have
-    *                    been consumed by nodes up to now.
-    *  @param context The context of what sub-transaction is being
-    *                 built.
-    *  @param aborted The error that lead to aborting the building of
-    *                 this transaction. We inline this error to allow
-    *                 reporting the error jointly with the state that
-    *                 the transaction was in when aborted. It is up to
-    *                 the caller to check for 'isAborted' after every
-    *                 change to a transaction.
-    *  @param keys A local store of the contract keys. Note that this contains
-    *              info both about relative and absolute contract ids. We must
-    *              do this because absolute contract ids can be archived as
-    *              part of execution, and we must record these archivals locally.
-    *              Note: it is important for keys that we know to not be present
-    *              to be present as [[None]]. The reason for this is that we must
-    *              record the "no key" information for absolute contract ids that
-    *              we archive. This is not an optimization and is required for
-    *              correct semantics, since otherwise lookups for keys for
-    *              locally archived absolute contract ids will succeed wrongly.
-    */
-  case class PartialTransaction(
-      submissionTime: Option[Time.Timestamp],
-      nextNodeIdx: Int,
-      nodes: HashMap[NodeId, Node],
-      consumedBy: Map[TContractId, NodeId],
-      context: PartialTransaction.Context,
-      aborted: Option[TransactionError],
-      keys: Map[GlobalKey, Option[TContractId]],
-  ) {
-
-    import PartialTransaction._
-
-    def nodesToString: String =
-      if (nodes.isEmpty) "<empty transaction>"
-      else {
-        val sb = new StringBuilder()
-
-        def addToStringBuilder(
-            nid: NodeId,
-            node: GenNode.WithTxValue[NodeId, TContractId],
-            rootPrefix: String,
-        ): Unit = {
-          sb.append(rootPrefix)
-            .append("node ")
-            .append(nid.index)
-            .append(": ")
-            .append(node.toString)
-            .append(", ")
-          ()
-        }
-
-        def removeTrailingComma(): Unit = {
-          if (sb.length >= 2) sb.setLength(sb.length - 2) // remove trailing ", "
-        }
-
-        // roots field is not initialized when this method is executed on a failed transaction,
-        // so we need to compute them.
-        val rootNodes = {
-          val allChildNodeIds: Set[NodeId] = nodes.values.flatMap {
-            case _: LeafOnlyNode[_, _] => Nil
-            case ex: NodeExercises[NodeId, _, _] => ex.children.toSeq
-          }(breakOut)
-
-          nodes.keySet diff allChildNodeIds
-        }
-        val tx = GenTransaction(nodes, ImmArray(rootNodes), None)
-
-        tx.foreach { (nid, node) =>
-          val rootPrefix = if (rootNodes.contains(nid)) "root " else ""
-          addToStringBuilder(nid, node, rootPrefix)
-        }
-        removeTrailingComma()
-
-        sb.toString
-      }
-
-    def resolveCidDiscriminator(node: Node) =
-      node.resolveRelCid(cid =>
-        Ref.ContractIdString.assertFromString("0" + cid.discriminator.get.toHexaString))
-
-    /** Finish building a transaction; i.e., try to extract a complete
-      *  transaction from the given 'PartialTransaction'. This fails if
-      *  the 'PartialTransaction' is not yet complete or has been
-      *  aborted.
-      */
-    def finish: Either[PartialTransaction, Transaction] =
-      context match {
-        case ContextRoot(transactionSeed, children) if aborted.isEmpty =>
-          Right(
-            GenTransaction(
-              nodes =
-                if (transactionSeed.isEmpty) nodes
-                else nodes.transform((_, v) => resolveCidDiscriminator(v)),
-              roots = children.toImmArray,
-              optUsedPackages = None,
-              transactionSeed = transactionSeed,
-            ),
-          )
-        case _ =>
-          Left(this)
-      }
-
-    /** Lookup the contract associated to 'RelativeContractId'.
-      * Return the contract instance and the node in which it was
-      * consumed if any.
-      */
-    def lookupLocalContract(
-        lcoid: RelativeContractId,
-    ): Option[(ContractInst[Transaction.Value[TContractId]], Option[NodeId])] =
-      for {
-        node <- nodes.get(lcoid.txnid)
-        coinst <- node match {
-          case create: NodeCreate.WithTxValue[TContractId] =>
-            Some((create.coinst, consumedBy.get(lcoid)))
-          case _: NodeExercises[_, _, _] | _: NodeFetch[_] | _: NodeLookupByKey[_, _] => None
-        }
-      } yield coinst
-
-    /** Extend the 'PartialTransaction' with a node for creating a
-      * contract instance.
-      */
-    def insertCreate(
-        coinst: ContractInst[Value[TContractId]],
-        optLocation: Option[Location],
-        signatories: Set[Party],
-        stakeholders: Set[Party],
-        key: Option[KeyWithMaintainers[Value[Nothing]]],
-    ): Either[String, (TContractId, PartialTransaction)] = {
-      val serializableErrs = serializable(coinst.arg)
-      if (serializableErrs.nonEmpty) {
-        Left(
-          s"""Trying to create a contract with a non-serializable value: ${serializableErrs.iterator
-            .mkString(",")}""",
-        )
-      } else {
-        val nodeSeed = deriveChildSeed
-        val nodeId = NodeId(nextNodeIdx)
-        val contractDiscriminator =
-          for {
-            seed <- nodeSeed
-            time <- submissionTime
-          } yield crypto.Hash.deriveContractDiscriminator(seed, time, stakeholders)
-        val cid = RelativeContractId(nodeId, contractDiscriminator)
-        val createNode = NodeCreate(
-          nodeSeed,
-          cid,
-          coinst,
-          optLocation,
-          signatories,
-          stakeholders,
-          key,
-        )
-        val ptx = copy(
-          nextNodeIdx = nextNodeIdx + 1,
-          context = context.addChild(nodeId),
-          nodes = nodes.updated(nodeId, createNode),
-        )
-
-        // if we have a contract key being added, include it in the list of
-        // active keys
-        key match {
-          case None => Right((cid, ptx))
-          case Some(kWithM) =>
-            val ck = GlobalKey(coinst.template, kWithM.key.value)
-            Right((cid, ptx.copy(keys = ptx.keys.updated(ck, Some(cid)))))
-        }
-      }
-    }
-
-    def serializable(a: Value[TContractId]): ImmArray[String] = a.value.serializable()
-
-    def insertFetch(
-        coid: TContractId,
-        templateId: TypeConName,
-        optLocation: Option[Location],
-        actingParties: Set[Party],
-        signatories: Set[Party],
-        stakeholders: Set[Party],
-    ): PartialTransaction =
-      mustBeActive(
-        coid,
-        templateId,
-        insertLeafNode(
-          NodeFetch(coid, templateId, optLocation, Some(actingParties), signatories, stakeholders),
-        ),
-      )
-
-    def insertLookup(
-        templateId: TypeConName,
-        optLocation: Option[Location],
-        key: KeyWithMaintainers[Value[Nothing]],
-        result: Option[TContractId],
-    ): PartialTransaction =
-      insertLeafNode(NodeLookupByKey(templateId, optLocation, key, result))
-
-    def beginExercises(
-        targetId: TContractId,
-        templateId: TypeConName,
-        choiceId: ChoiceName,
-        optLocation: Option[Location],
-        consuming: Boolean,
-        actingParties: Set[Party],
-        signatories: Set[Party],
-        stakeholders: Set[Party],
-        controllers: Set[Party],
-        mbKey: Option[KeyWithMaintainers[Value[Nothing]]],
-        chosenValue: Value[TContractId],
-    ): Either[String, PartialTransaction] = {
-      val serializableErrs = serializable(chosenValue)
-      if (serializableErrs.nonEmpty) {
-        Left(
-          s"""Trying to exercise a choice with a non-serializable value: ${serializableErrs.iterator
-            .mkString(",")}""",
-        )
-      } else {
-        val nodeId = NodeId(nextNodeIdx)
-        Right(
-          mustBeActive(
-            targetId,
-            templateId,
-            copy(
-              nextNodeIdx = nextNodeIdx + 1,
-              context = ContextExercise(
-                ExercisesContext(
-                  contextSeed = deriveChildSeed,
-                  targetId = targetId,
-                  templateId = templateId,
-                  contractKey = mbKey,
-                  choiceId = choiceId,
-                  optLocation = optLocation,
-                  consuming = consuming,
-                  actingParties = actingParties,
-                  chosenValue = chosenValue,
-                  signatories = signatories,
-                  stakeholders = stakeholders,
-                  controllers = controllers,
-                  nodeId = nodeId,
-                  parent = context,
-                ),
-              ),
-              // important: the semantics of DAML dictate that contracts are immediately
-              // inactive as soon as you exercise it. therefore, mark it as consumed now.
-              consumedBy = if (consuming) consumedBy.updated(targetId, nodeId) else consumedBy,
-              keys = mbKey match {
-                case Some(kWithM) if consuming =>
-                  keys.updated(GlobalKey(templateId, kWithM.key.value), None)
-                case _ => keys
-              },
-            ),
-          ),
-        )
-      }
-    }
-
-    def endExercises(value: Value[TContractId]): PartialTransaction =
-      context match {
-        case ContextExercise(ec, children) =>
-          val exerciseNode: Transaction.Node = NodeExercises(
-            nodeSeed = ec.contextSeed,
-            targetCoid = ec.targetId,
-            templateId = ec.templateId,
-            choiceId = ec.choiceId,
-            optLocation = ec.optLocation,
-            consuming = ec.consuming,
-            actingParties = ec.actingParties,
-            chosenValue = ec.chosenValue,
-            stakeholders = ec.stakeholders,
-            signatories = ec.signatories,
-            controllers = ec.controllers,
-            children = children.toImmArray,
-            exerciseResult = Some(value),
-            key = ec.contractKey,
-          )
-          val nodeId = ec.nodeId
-          copy(context = ec.parent.addChild(nodeId), nodes = nodes.updated(nodeId, exerciseNode))
-        case ContextRoot(_, _) =>
-          noteAbort(EndExerciseInRootContext)
-      }
-
-    /** Note that the transaction building failed due to the given error */
-    def noteAbort(err: TransactionError): PartialTransaction = copy(aborted = Some(err))
-
-    /** `True` iff the given `ContractId` has been consumed already */
-    def isConsumed(coid: TContractId): Boolean = consumedBy.contains(coid)
-
-    /** Guard the execution of a step with the unconsumedness of a
-      * `ContractId`
-      */
-    def mustBeActive(
-        coid: TContractId,
-        templateId: TypeConName,
-        f: => PartialTransaction,
-    ): PartialTransaction =
-      consumedBy.get(coid) match {
-        case None => f
-        case Some(nid) => noteAbort(ContractNotActive(coid, templateId, nid))
-      }
-
-    /** Insert the given `LeafNode` under a fresh node-id, and return it */
-    def insertLeafNode(node: LeafNode): PartialTransaction = {
-      val nodeId = NodeId(nextNodeIdx)
-      copy(
-        nextNodeIdx = nextNodeIdx + 1,
-        context = context.addChild(nodeId),
-        nodes = nodes.updated(nodeId, node),
-      )
-    }
-
-    def deriveChildSeed: Option[crypto.Hash] =
-      context.contextSeed.map(crypto.Hash.deriveNodeSeed(_, nodes.size))
-
-  }
 
 }
